requirements.txt
Flask==3.1.0
python-telegram-bot==13.7
gunicorn==23.0.0
requests==2.32.3
